+++ src/project-root.ts
@@ -134,6 +134,17 @@ export async function pickProjectRoot(
     return rootToUri(uris[0]);
   }
 
+  uris.sort((a, b) => {
+    if (!selected) {
+      return 0;
+    }
+    return rootToUri(a).fsPath === selected.fsPath
+      ? -1
+      : rootToUri(b).fsPath === selected.fsPath
+      ? 1
+      : 0;
+  });
+
   const project_root_options = uris.map((root_or_uri) => {
     let uri;
     let reason;